requirements.txt
streamlit>=1.56.0
openai>=2.32.0
python-dotenv>=1.0.0